+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Fix using node as offset in position mixin
 - Fix using Height Match component on elements with different offsetParents
 - Fix `scope` command for already scoped css, comments are removed from generated css files
+- Fix Sticky placeholder height on resize
 
 ## 3.0.0 rc 17 (September 27, 2018)